some new changes

+++ source/ft_strlen.c
@@ -13,7 +13,7 @@
 #include <stdio.h>
 #include <string.h>
 
-unsigned int	ft_strlen(const char *a)
+size_t	ft_strlen(const char *a)
 {
 	int	i;
